Adding Beta and Preview products to main artifact
Previously, Beta and Preview products were only included in the `alpha`
artifact. They are now being included in the main artifact to ease
product discoverability and the collective operational overhead of
maintaining multiple artifacts per library.

If you have questions or concerns, raise them in a Github issue and we
will be sure to address.